tipc: add new tipc configuration tool
tipc is a user-space configuration tool for TIPC (Transparent
Inter-process Communication). It utilizes the TIPC netlink API in the
kernel to fetch data or perform actions.

The tipc tool has somewhat similar syntax to the ip tool meaning that
users of the ip tools should not feel that unfamiliar with this tool.
The internal design of the tipc tool is however a bit different from
the ip tool. See the README file for more information about the design
decisions.